[FLINK-26463][table-planner] Use MiniCluster for TableEnvironmentITCase and moved some tests to TableEnvironmentTest
This closes #18968.
